Revisioni e ricambi sql ok!

+++ sql/ddl/table/ricambio.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `tripdb`.`ricambio` (
   `Codice` varchar(45) NOT NULL,
-  `CostoUnitario` int unsigned NOT NULL,
+  `CostoUnitario` decimal(8,2) NOT NULL,
   `QuantitàDiRiordino` int unsigned NOT NULL,
   `Descrizione` varchar(1000) NOT NULL,
   `ScortaMinima` int unsigned NOT NULL,
